feature: add signin method
