+++ frontend/src/pages/Home/Login/styles.js
@@ -1,20 +1,9 @@
 import { makeStyles } from '@material-ui/core/styles';
-import { imageBg } from '../../assets/img/index';
 
 export const useStyles = makeStyles((theme) => ({
-  root: {
-    height: '100vh',
-  },
-  image: {
-    backgroundImage: `url(${imageBg})`,
-    backgroundRepeat: 'no-repeat',
-    backgroundSize: 'cover',
-    backgroundPosition: 'center',
-    width: '100vw',
-    height: '100vh',
-  },
   paper: {
     margin: theme.spacing(8, 4),
+    marginBottom: 4,
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center',
@@ -27,7 +16,7 @@ export const useStyles = makeStyles((theme) => ({
     margin: theme.spacing(3, 0, 2),
   },
   containerBtnLoader: {
-    margin: theme.spacing(1),
+    margin: theme.spacing(0),
     position: 'relative',
   },
 }));
