+++ frontend/src/routes/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, BrowserRouter, Route } from 'react-router-dom';
 
-import { Login } from '../pages/Login';
+import { Home } from '../pages/Home';
 import { Perfil } from '../pages/Perfil';
 import { NavBar } from '../components/NavBar';
 
@@ -9,7 +9,7 @@ export const Routes = () => (
   <BrowserRouter>
     <Route path="/dashboard" component={NavBar} />
     <Switch>
-      <Route path="/" exact component={Login} />
+      <Route path="/" exact component={Home} />
       <Route path="/dashboard/perfil" exact component={Perfil} />
     </Switch>
   </BrowserRouter>
